Reorganize shell namespaces
Closes issue #67
