+++ dolang-ext-shell/tests/mod/agent.dol
@@ -7,8 +7,8 @@ import
 let agent_bin = nil
 let Vfs = nil
 
-if (sys.os != "windows")
-  import container:
+if (sys.os().archetype != :windows:)
+  import shell:
     Vfs: it
   Vfs = it
   agent_bin = env.get DOLANG_SHELL_VFS
@@ -30,6 +30,6 @@ pub def with_agent block
 pub def with_and_without_agent test block
   log "$test: without agent"
   block()
-  if (sys.os != "windows")
+  if (sys.os().archetype != :windows:)
       log "$test: with agent"
       with_agent $block
