+++ dolang-ext-shell/tests/regression/env.dol
@@ -2,7 +2,7 @@ import sys
 import std:
   - MissingPosError
 
-if (sys.os == "windows")
+if (sys.os().archetype == :windows:)
   log "skipped on Windows"
   return
