+++ dolang-ext-shell/tests/regression/fs/chown.dol
@@ -1,6 +1,6 @@
 import sys
 
-if (sys.os == "windows")
+if (sys.os().archetype == :windows:)
   log "skipped on Windows"
   return
