+++ dolang-ext-shell/tests/regression/fs/utime.dol
@@ -23,7 +23,7 @@ with_and_without_agent "set_timestamps updates timestamps" do
   open $test_path w do |f|
     f.write "set_timestamps test"
 
-  if (sys.os == "windows")
+  if (sys.os().archetype == :windows:)
     if (env.get "WINEPREFIX")
       log "skipping Windows created timestamp test under wine"
       remove $test_path
