+++ dolang-ext-shell/tests/regression/pipe_torture.dol
@@ -4,7 +4,7 @@ import strand:
 import time:
   - sleep
 
-if (sys.os == "windows")
+if (sys.os().archetype == :windows:)
   log "not applicable"
   return
