+++ dolang-shell/lib/container/docker.dol
@@ -1,8 +1,6 @@
 import
   proc.run:
     docker: docker_cli
-  sys:
-    - env
   container._dockman: dockman
 
 pub let Image = dockman.Image
