+++ dolang-shell/lib/sudo.dol
@@ -1,5 +1,5 @@
 import xdg
-  container:
+  shell:
     - Vfs
   container._common:
     - agent_bin
